requirements.txt
flask
pypdf
python-dotenv

# Core LangChain
langchain
langchain-community
langchain-openai
langchain-experimental

# Pinecone vector DB
# pinecone-client[grpc]
pinecone
langchain-pinecone
langchain_huggingface
langgraph

# Optional: Hugging Face support
huggingface-hub
transformers

# For acceleration (typo fix)
accelerate

# Auth0
authlib
requests

# Additional libraries
flask-cors  # For handling CORS in Flask
gunicorn    # For deploying Flask apps
openai      # For OpenAI API integration
numpy       # For numerical computations
pandas      # For data manipulation
scikit-learn  # For machine learning tasks
matplotlib  # For data visualization
seaborn     # For statistical data visualization
tqdm        # For progress bars
joblib
pytorch_lightning
#python3 -m pip install -r requirements.txt